requirements.txt
PyMuPDF
pandas
python-dateutil